pyproject.toml
[tool.poetry]
name = "epub-compressor"
version = "0.1.0"
description = "Compress epub by converting images to webp"
authors = ["icealtria <icealtria@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.3.0"
pillow-avif-plugin = "^1.4.3"
streamlit = "^1.35.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
epub-compressor = "epub_compressor.cli:main"